+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## Changed
+- Type declarations added to `TypeInterface` (#129)
+
 ## Removed
 - Compatibility with TYPO3 v11 (#100)
 - Compatibility with TYPO3 v12 (#119)